added current weather

+++ app/src/main/java/com/taru/data/remote/weather/dto/inner/WeatherLocationDto.kt
@@ -6,5 +6,5 @@ import com.squareup.moshi.JsonClass
  * Created by Niraj on 17-01-2023.
  */
 @JsonClass(generateAdapter = true)
-data class WeatherLocationDto(val lon: Double, val lat: Double) {
+data class WeatherLocationDto(val lon: Float, val lat: Float) {
 }
