+++ app/src/main/java/com/taru/domain/weather/enitity/ModelWeather.kt
@@ -6,5 +6,5 @@ import com.squareup.moshi.JsonClass
  * Created by Niraj on 16-01-2023.
  */
 
-data class ModelWeather(val lat: Double, val lng: Double) {
+data class ModelWeather(val lat: Float, val lng: Float) {
 }
